优化代码架构
