+++ everyday_wechat/control/__init__.py
@@ -1,0 +1,6 @@
+#! usr/bin/env python
+# -*- coding: utf-8 -*-
+# Date: 2019/6/23
+# Author: snow
+
+
