+++ everyday_wechat/control/onewords/hitokoto.py
