+++ everyday_wechat/utils/data_collection.py
@@ -3,15 +3,11 @@
 import importlib
 from datetime import datetime
 
-from weather.sojson import get_today_weather
-# from weather.rtweather import get_rttodayweather
-# from bot.qingyunke import get_qingyunke
-# from bot.tuling123 import get_tuling123
-# from bot.yigeai import get_yigeai
-from main.common import (
+from everyday_wechat.control.weather.rtweather import get_today_weather
+from everyday_wechat.utils.common import (
     get_yaml
 )
-from horoscope.spider_script import get_constellation, get_xzw_data_list
+from everyday_wechat.control.horoscope.spider_script import get_constellation, get_xzw_data_list
 
 DICTUM_NAME_DICT = {1: 'wufazhuce', 2: 'acib', 3: 'lovelive', 4: 'hitokoto', 5: 'rtjokes', 6: 'scapy'}
 BOT_NAME_DICT = {1: 'tuling123', 2: 'yigeai', 3: 'qingyunke'}
@@ -26,7 +22,7 @@ def get_dictum_info(channel):
         return None
     source = DICTUM_NAME_DICT.get(channel, '')
     if source:
-        addon = importlib.import_module('onewords.' + source, __package__)
+        addon = importlib.import_module('everyday_wechat.control.onewords.' + source, __package__)
         dictum = addon.get_one_words()
         # print(dictum)
         return dictum
@@ -55,7 +51,7 @@ def get_bot_info(message, userId=''):
     channel = get_yaml().get('bot_channel', 3)
     source = BOT_NAME_DICT.get(channel, 'qingyunke')
     if source:
-        addon = importlib.import_module('bot.' + source, __package__)
+        addon = importlib.import_module('everyday_wechat.control.bot.' + source, __package__)
         reply_msg = addon.get_auto_reply(message, userId)
         return reply_msg
     # reply_msg = get_tuling123(message)
